Site is now aware of series and repo attributes
The series attribute allows posts of the same series to collate
together. At the beginning of the post, a blockquote is placed which
identifies the post as part of a series. Following that, a list of posts
in that series is printed (in reverse chronological order) with links to
each part. The current part is unlinked and serves to identify where the
reader is.

The repo attribute is just the location of the repository relating to
the code discussed in the blog. This is combined with the series list,
if that is present.